\name{aircraft.data}
\alias{aircraft.data}
\docType{data}
\title{
THE AIRCRAFT DATA SET
}
\description{
See Understanding Biplots.
}
\usage{data(aircraft.data)}
\format{
  A data frame with 21 observations on the following 5 variables.
  \describe{
    \item{\code{Aircraft}}{a factor with levels \code{F-100A} \code{F-101A} \code{F-102A} \code{F-104A} \code{F-105B} \code{F-106A} \code{F-4B} \code{F-86A} \code{F-89A} \code{F-8A} \code{F-94A} \code{F11F-1} \code{F3D-1} \code{F3H-2} \code{F4D-1} \code{F9F-2} \code{F9F-6} \code{FH-1} \code{FJ-1} \code{XF10F-1} \code{YF-107A}}
    \item{\code{SPR}}{a numeric vector}
    \item{\code{RGF}}{a numeric vector}
    \item{\code{PLF}}{a numeric vector}
    \item{\code{SLF}}{a numeric vector}
  }
}

\keyword{datasets}
