\name{archaeology.data}
\alias{archaeology.data}
\docType{data}
\title{
THE ARCHAEOLOGY DATA SET
}
\description{
See Understanding Biplots
}
\usage{data(archaeology.data)}
\format{
  A data frame with 711 observations on the following 8 variables.
  \describe{
    \item{\code{platfang}}{a numeric vector}
    \item{\code{platfwth}}{a numeric vector}
    \item{\code{platfthk}}{a numeric vector}
    \item{\code{pointlng}}{a numeric vector}
    \item{\code{pointwth}}{a numeric vector}
    \item{\code{pointthk}}{a numeric vector}
    \item{\code{Site}}{a factor with levels \code{MSAI} \code{MSAIILOWER} \code{MSAIIUPPER}}
    \item{\code{ratio}}{a numeric vector}
  }
}

\keyword{datasets}
