\name{CopperFroth.data}
\alias{CopperFroth.data}
\docType{data}
\title{
THE COPPER FROTH DATA SET
}
\description{
See Understanding Biplots
}
\usage{data(CopperFroth.data)}
\format{
  A data frame with 497 observations on the following 9 variables.
  \describe{
    \item{\code{X1}}{a numeric vector}
    \item{\code{X2}}{a numeric vector}
    \item{\code{X3}}{a numeric vector}
    \item{\code{X4}}{a numeric vector}
    \item{\code{X5}}{a numeric vector}
    \item{\code{X6}}{a numeric vector}
    \item{\code{X7}}{a numeric vector}
    \item{\code{X8}}{a numeric vector}
    \item{\code{CLASS}}{a numeric vector}
  }
}


\keyword{datasets}
