\name{CVABlGeog.data}
\alias{CVABlGeog.data}
\docType{data}
\title{
A SUBSET OF THE EDUCATION DATA SET
}
\description{
See Understanding Biplots
}
\usage{data(CVABlGeog.data)}
\format{
  A data frame with 605 observations on the following 6 variables.
  \describe{
    \item{\code{TOTScore}}{a numeric vector}
    \item{\code{eduyrs}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{apcexdec}}{a numeric vector}
    \item{\code{mEdY}}{a numeric vector}
    \item{\code{groepe}}{a factor with levels \code{Metro} \code{Rural} \code{Urban}}
  }
}

\keyword{datasets}
