\name{CVARace.data}
\alias{CVARace.data}
\docType{data}
\title{
A SUBSET OF THE EDUCATION DATA SET
}
\description{
See Understanding Biplots
}
\usage{data(CVARace.data)}
\format{
  A data frame with 799 observations on the following 6 variables.
  \describe{
    \item{\code{TOTScore}}{a numeric vector}
    \item{\code{eduyrs}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{pcexpdec}}{a numeric vector}
    \item{\code{mEdY}}{a numeric vector}
    \item{\code{groepe}}{a factor with levels \code{Black} \code{Coloured} \code{Indian} \code{White}}
  }
}

\keyword{datasets}
