\name{Education93.data}
\alias{Education93.data}
\docType{data}
\title{
THE 1993 EDUCATION DATA SET
}
\description{
See Understanding Biplots
}
\usage{data(Education93.data)}
\format{
  A data frame with 799 observations on the following 6 variables.
  \describe{
    \item{\code{TScore}}{a numeric vector}
    \item{\code{EducY}}{a numeric vector}
    \item{\code{Age}}{a numeric vector}
    \item{\code{PcExpD}}{a numeric vector}
    \item{\code{ParEdY}}{a numeric vector}
    \item{\code{Group}}{a factor with levels \code{Black} \code{Coloured} \code{Indian} \code{White}}
  }
}


\keyword{datasets}
