\name{Fig.3.19.dat}
\alias{Fig.3.19.dat}
\docType{data}
\title{
THE FIG 3.19. DATA SET
}
\description{
See Understanding Biplots
}
\usage{data(Fig.3.19.dat)}
\format{
  The format is:
 num [1:100, 1:3] 13.01 19.47 21.79 9.47 14 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : NULL
}

\keyword{datasets}
