\name{Flotation.data}
\alias{Flotation.data}
\docType{data}
\title{
THE FLOTATION DATA SET
}
\description{
See Understanding Biplots
}
\usage{data(Flotation.data)}
\format{
  A data frame with 1425 observations on the following 7 variables.
  \describe{
    \item{\code{RECORD}}{a numeric vector}
    \item{\code{RMOF}}{a numeric vector}
    \item{\code{RMP}}{a numeric vector}
    \item{\code{BMSWA}}{a numeric vector}
    \item{\code{BMCFF}}{a numeric vector}
    \item{\code{BMCFD}}{a numeric vector}
    \item{\code{EXFFBB}}{a numeric vector}
  }
}

\keyword{datasets}
