\name{Funding.data}
\alias{Funding.data}
\docType{data}
\title{
THE FUNDING DATA SET
}
\description{
See Understanding Biplots
}
\usage{data(Funding.data)}
\format{
  A data frame with 10 observations on the following 5 variables.
  \describe{
    \item{\code{A}}{a numeric vector}
    \item{\code{B}}{a numeric vector}
    \item{\code{C}}{a numeric vector}
    \item{\code{D}}{a numeric vector}
    \item{\code{E}}{a numeric vector}
  }
}

\keyword{datasets}
