\name{Headdimensions.data}
\alias{Headdimensions.data}
\docType{data}
\title{
THE HEAD DIMENSIONS DATA SET
}
\description{
See Understanding Biplots
}
\usage{data(Headdimensions.data)}
\format{
  A data frame with 259 observations on the following 7 variables.
  \describe{
    \item{\code{Sex}}{a factor with levels \code{F} \code{M}}
    \item{\code{MFB}}{a numeric vector}
    \item{\code{BAM}}{a numeric vector}
    \item{\code{TFH}}{a numeric vector}
    \item{\code{LGAN}}{a numeric vector}
    \item{\code{LTN}}{a numeric vector}
    \item{\code{LTG}}{a numeric vector}
  }
}

\keyword{datasets}
