\name{LouiseRika13tot18.data}
\alias{LouiseRika13tot18.data}
\docType{data}
\title{
LOUISERIKA13to18 data
}
\description{
Subset of LouiseRika.data
}
\usage{data(LouiseRika13tot18.data)}
\format{
  A data frame with 10410 observations on the following 8 variables.
  \describe{
    \item{\code{TScore}}{a numeric vector}
    \item{\code{EducY}}{a numeric vector}
    \item{\code{Age}}{a numeric vector}
    \item{\code{PcExpD}}{a numeric vector}
    \item{\code{ParEdY}}{a numeric vector}
    \item{\code{Group}}{a factor with levels \code{Black} \code{Colrd} \code{Indian} \code{White}}
    \item{\code{Year}}{a numeric vector}
    \item{\code{Groups.8}}{a factor with levels \code{Black1993} \code{Black2006} \code{Colrd1993} \code{Colrd2006} \code{Indian1993} \code{Indian2006} \code{White1993} \code{White2006}}
  }
}
