\name{MCA.Table.1.data.ord}
\alias{MCA.Table.1.data.ord}
\docType{data}
\title{
Data for 
}
\description{
Ordinal form of dta in Table 8.1 of Understanding Biplots
}
\usage{data(MCA.Table.1.data.ord)}
\format{
  A data frame with 7 observations on the following 4 variables.
  \describe{
    \item{\code{ID}}{a factor with levels \code{Alisdair} \code{George} \code{Harriet} \code{Ivor} \code{Jane} \code{Jeremy} \code{Myfanwy}}
    \item{\code{Gender}}{a factor with levels \code{F} \code{M}}
    \item{\code{Hair}}{an ordered factor with levels \code{Grey} < \code{Fair} < \code{Brown} < \code{Dark}}
    \item{\code{Education}}{an ordered factor with levels \code{School} < \code{University} < \code{Postgrad}}
  }
}
