\name{MCA.Table.1.data}
\alias{MCA.Table.1.data}
\docType{data}
\title{
Table 8.1 of Understanding Biplots
}
\description{
See Understanding Biplots Chapter 8.
}
\usage{data(MCA.Table.1.data)}
\format{
  A data frame with 7 observations on the following 6 variables.
  \describe{
    \item{\code{ID}}{a factor with levels \code{Alisdair} \code{George} \code{Harriet} \code{Ivor} \code{Jane} \code{Jeremy} \code{Myfanwy}}
    \item{\code{Gender}}{a factor with levels \code{F} \code{M}}
    \item{\code{Hair}}{an ordered factor with levels \code{Fair} < \code{Grey} < \code{Brown} < \code{Dark}}
    \item{\code{Region}}{a factor with levels \code{England} \code{Scotland} \code{Wales}}
    \item{\code{Work}}{an ordered factor with levels \code{Manual} < \code{Clerical} < \code{Professional}}
    \item{\code{Education}}{an ordered factor with levels \code{School} < \code{University} < \code{Postgrad}}
  }
}
