\name{Mondi.StrengthPlot.data}
\alias{Mondi.StrengthPlot.data}
\docType{data}
\title{
PAPER QUALITIES
}
\description{
Data set containing information on paper qualities
}
\usage{data(Mondi.StrengthPlot.data)}
\format{
  A data frame with 290 observations on the following 6 variables.
  \describe{
    \item{\code{Tree}}{a factor with levels \code{A} \code{B} \code{C} \code{D} \code{E}}
    \item{\code{Class}}{a factor with levels \code{A0} \code{A1} \code{A3} \code{A4} \code{A5} \code{B0} \code{B1} \code{B2} \code{B3} \code{B4} \code{B5} \code{C0} \code{C1} \code{C2} \code{C3} \code{C4} \code{C5} \code{D0} \code{D1} \code{D2} \code{D3} \code{D4} \code{D5} \code{E0} \code{E1} \code{E2} \code{E3} \code{E4} \code{E5}}
    \item{\code{BurstIndex}}{a numeric vector}
    \item{\code{TearIndex}}{a numeric vector}
    \item{\code{BreakLen}}{a numeric vector}
    \item{\code{plotcol}}{a numeric vector}
  }
}
