\name{my.colours}
\alias{my.colours}
\docType{data}
\title{
SPECIFY A COLOUR SET
}
\description{
Colour set specification
}
\usage{data(my.colours)}
\format{
  The format is:
 chr [1:12] "#000000" "#A0A0A0" "#808080" "#FF0000" "#00FF00" "#0000FF" 
 "#FF8000" "#FF00FF" "#808000" "#FFFF00" "#00FFFF" "#7DBEFF"
}
