\name{Ocotea.cat}
\alias{Ocotea.cat}
\docType{data}
\title{
CATEGORIZED VERSION OF OCOTEA DATA
}
\description{
See Understanding Biplots
}
\usage{data(Ocotea.cat)}
\format{
  A data frame with 37 observations on the following 8 variables.
  \describe{
    \item{\code{Sample}}{a numeric vector}
    \item{\code{Species}}{a factor with levels \code{Obul} \code{Oken} \code{Opor}}
    \item{\code{VesD}}{a factor with levels \code{VD1} \code{VD2} \code{VD3} \code{VD4} \code{VD5}}
    \item{\code{VesL}}{a factor with levels \code{VL1} \code{VL2} \code{VL3} \code{VL4} \code{VL5}}
    \item{\code{FibL}}{a factor with levels \code{F1} \code{F2} \code{F3} \code{F4} \code{F5}}
    \item{\code{RayH}}{a factor with levels \code{RH1} \code{RH2} \code{RH3} \code{RH4} \code{RH5}}
    \item{\code{RayW}}{a factor with levels \code{RW1} \code{RW2} \code{RW3} \code{RW4} \code{RW5}}
    \item{\code{NumVes}}{a factor with levels \code{N1} \code{N2} \code{N3} \code{N4} \code{N5}}
  }
}
