\name{Ocotea.data}
\alias{Ocotea.data}
\docType{data}
\title{
THE OCOTEA DATA SET
}
\description{
See Understanding Biplots
}
\usage{data(Ocotea.data)}
\format{
  A data frame with 37 observations on the following 8 variables.
  \describe{
    \item{\code{Sample}}{a numeric vector}
    \item{\code{Species}}{a factor with levels \code{Obul} \code{Oken} \code{Opor}}
    \item{\code{VesD}}{a numeric vector}
    \item{\code{VesL}}{a numeric vector}
    \item{\code{FibL}}{a numeric vector}
    \item{\code{RayH}}{a numeric vector}
    \item{\code{RayW}}{a numeric vector}
    \item{\code{NumVes}}{a numeric vector}
  }
}

\keyword{datasets}
