\name{palette.default.colours}
\alias{palette.default.colours}
\docType{data}
\title{
DEFAULT COLOUR PALETTE
}
\description{
A default colour palette.
}
\usage{data(palette.default.colours)}
\format{
  The format is:
 chr [1:8] "black" "red" "green3" "blue" "cyan" "magenta" ...
}

\keyword{datasets}
