\name{PCA.ex.data}
\alias{PCA.ex.data}
\docType{data}
\title{
PCA EXAMPLE DATA SET
}
\description{
See Understanding Biplots
}
\usage{data(PCA.ex.data)}
\format{
  The format is:
 num [1:25, 1:3] 13.5 15.3 17.8 18.4 18 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:3] "X" "Y" "Z"
}

\keyword{datasets}
