\name{Pine.data}
\alias{Pine.data}
\docType{data}
\title{
THE PINUS DATA SET
}
\description{
See Understanding Biplots
}
\usage{data(Pine.data)}
\format{
  A data frame with 36 observations on the following 9 variables.
  \describe{
    \item{\code{species}}{a factor with levels \code{P.ell} \code{P.kes} \code{P.max} \code{P.pat} \code{P.tae}}
    \item{\code{TotYield}}{a numeric vector}
    \item{\code{Alkali}}{a numeric vector}
    \item{\code{Density}}{a numeric vector}
    \item{\code{TEA}}{a numeric vector}
    \item{\code{Tensile}}{a numeric vector}
    \item{\code{Tear}}{a numeric vector}
    \item{\code{Burst}}{a numeric vector}
    \item{\code{Growth}}{a numeric vector}
  }
}

\keyword{datasets}
