\name{pred.regions.cat}
\alias{pred.regions.cat}
\title{
pred.regions.cat
}
\description{
Use for constructing prediction regions for categorical data. Called by AOD.cat.predregions)
}
\usage{
pred.regions.cat(Z.rj, cij2, var.j, x.grid = 0.05, y.grid = 0.05, 
plot.symbol = 15, size = 0.25, colours = 1:10)
}
