\name{ProcessQuality.data}
\alias{ProcessQuality.data}
\docType{data}
\title{
THE PROCESS QUALITY DATA SET
}
\description{
See Understanding Biplots
}
\usage{data(ProcessQuality.data)}
\format{
  The format is:
 num [1:16, 1:15] 50 52.1 51.8 51.3 51.2 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:16] "Target" "Jan00" "Feb00" "Mar00" ...
  ..$ : chr [1:15] "A1" "A2" "A3" "A4" ...
}

\keyword{datasets}
