\name{ReactionKinetic.data}
\alias{ReactionKinetic.data}
\docType{data}
\title{
THE REACTION KINETIC DATA SET
}
\description{
See Understanding Biplots
}
\usage{data(ReactionKinetic.data)}
\format{
  The format is:
 num [1:24, 1:5] 0 0.043 0.0687 0.0756 0.0789 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:24] "1" "2" "3" "4" ...
  ..$ : chr [1:5] "y" "x1" "x2" "x3" ...
}

\keyword{datasets}
