\name{Remuneration.cat.data.2002}
\alias{Remuneration.cat.data.2002}
\docType{data}
\title{
THE 2002 REMUNERATION DATA SET IN CATEGORICAL FORM
}
\description{
See Understanding Biplots
}
\usage{data(Remuneration.cat.data.2002)}
\format{
  A data frame with 728 observations on the following 11 variables.
  \describe{
    \item{\code{ID}}{a factor with levels \code{1} \code{10} \code{100} \code{1001} \code{101} \code{1010} \code{1011} \code{102} \code{1021} \code{103} \code{1031} \code{104} \code{1041} \code{105} \code{1051} \code{106} \code{1061} \code{107} \code{1071} \code{108} \code{1081} \code{109} \code{1091} \code{11} \code{110} \code{1100} \code{1101} \code{111} \code{1110} \code{1111} \code{112} \code{1121} \code{113} \code{1131} \code{114} \code{1141} \code{115} \code{1151} \code{116} \code{1161} \code{117} \code{1171} \code{118} \code{1181} \code{119} \code{1191} \code{12} \code{120} \code{1201} \code{121} \code{1210} \code{1211} \code{122} \code{1221} \code{123} \code{1231} \code{124} \code{1241} \code{125} \code{1251} \code{126} \code{1261} \code{127} \code{1271} \code{128} \code{1281} \code{129} \code{1291} \code{13} \code{130} \code{1301} \code{131} \code{1310} \code{1311} \code{132} \code{1321} \code{133} \code{1331} \code{134} \code{1341} \code{135} \code{1351} \code{136} \code{1361} \code{137} \code{1371} \code{138} \code{1381} \code{139} \code{1391} \code{14} \code{140} \code{1401} \code{141} \code{1410} \code{1411} \code{142} \code{1421} \code{143} \code{1431} \code{144} \code{1441} \code{145} \code{1451} \code{146} \code{1461} \code{147} \code{1471} \code{148} \code{1481} \code{149} \code{1491} \code{15} \code{150} \code{1501} \code{151} \code{1510} \code{1511} \code{152} \code{1521} \code{153} \code{1531} \code{154} \code{1541} \code{155} \code{1551} \code{156} \code{1561} \code{157} \code{1571} \code{158} \code{1581} \code{159} \code{1591} \code{16} \code{160} \code{1601} \code{161} \code{1610} \code{1611} \code{162} \code{1621} \code{163} \code{1631} \code{164} \code{1641} \code{165} \code{1651} \code{166} \code{1661} \code{167} \code{1671} \code{168} \code{1681} \code{169} \code{1691} \code{17} \code{170} \code{1701} \code{171} \code{1710} \code{1711} \code{172} \code{1721} \code{173} \code{1731} \code{174} \code{1741} \code{175} \code{1751} \code{176} \code{1761} \code{177} \code{1771} \code{178} \code{1781} \code{179} \code{1791} \code{18} \code{180} \code{1801} \code{181} \code{1810} \code{1811} \code{182} \code{1821} \code{183} \code{1831} \code{184} \code{1841} \code{185} \code{1851} \code{186} \code{1861} \code{187} \code{1871} \code{188} \code{1881} \code{189} \code{1891} \code{19} \code{190} \code{1901} \code{191} \code{1910} \code{1911} \code{192} \code{1921} \code{193} \code{1931} \code{194} \code{1941} \code{195} \code{1951} \code{196} \code{1961} \code{197} \code{1971} \code{198} \code{1981} \code{199} \code{1991} \code{2} \code{20} \code{200} \code{2001} \code{201} \code{2010} \code{2011} \code{202} \code{2021} \code{203} \code{2031} \code{204} \code{2041} \code{205} \code{2051} \code{206} \code{2061} \code{207} \code{2071} \code{208} \code{2081} \code{209} \code{2091} \code{21} \code{210} \code{2100} \code{2101} \code{211} \code{2110} \code{2111} \code{212} \code{2121} \code{213} \code{2131} \code{214} \code{2141} \code{215} \code{2151} \code{216} \code{2161} \code{217} \code{2171} \code{218} \code{2181} \code{219} \code{2191} \code{22} \code{220} \code{2201} \code{221} \code{2210} \code{2211} \code{222} \code{2221} \code{223} \code{2231} \code{224} \code{2241} \code{225} \code{2251} \code{226} \code{2261} \code{227} \code{2271} \code{228} \code{2281} \code{229} \code{2291} \code{23} \code{230} \code{2301} \code{231} \code{2310} \code{2311} \code{232} \code{2321} \code{233} \code{2331} \code{234} \code{2341} \code{235} \code{2351} \code{236} \code{2361} \code{237} \code{2371} \code{238} \code{2381} \code{239} \code{2391} \code{24} \code{240} \code{2401} \code{241} \code{2410} \code{2411} \code{242} \code{2421} \code{243} \code{2431} \code{244} \code{2441} \code{245} \code{2451} \code{246} \code{2461} \code{247} \code{2471} \code{248} \code{2481} \code{249} \code{2491} \code{25} \code{250} \code{2501} \code{251} \code{2510} \code{2511} \code{252} \code{2521} \code{253} \code{2531} \code{254} \code{2541} \code{255} \code{2551} \code{256} \code{2561} \code{257} \code{2571} \code{258} \code{2581} \code{259} \code{2591} \code{26} \code{260} \code{2601} \code{261} \code{2610} \code{2611} \code{262} \code{2621} \code{263} \code{2631} \code{264} \code{2641} \code{265} \code{2651} \code{266} \code{2661} \code{267} \code{2671} \code{268} \code{2681} \code{269} \code{2691} \code{27} \code{270} \code{2701} \code{271} \code{2710} \code{2711} \code{272} \code{2721} \code{273} \code{2731} \code{274} \code{2741} \code{275} \code{2751} \code{276} \code{2761} \code{277} \code{2771} \code{278} \code{2781} \code{279} \code{2791} \code{28} \code{280} \code{2801} \code{281} \code{2810} \code{2811} \code{282} \code{2821} \code{283} \code{2831} \code{284} \code{2841} \code{285} \code{2851} \code{286} \code{2861} \code{287} \code{2871} \code{288} \code{2881} \code{289} \code{2891} \code{29} \code{290} \code{2901} \code{291} \code{2910} \code{2911} \code{292} \code{2921} \code{293} \code{2931} \code{294} \code{2941} \code{295} \code{2951} \code{296} \code{2961} \code{297} \code{2971} \code{298} \code{2981} \code{299} \code{2991} \code{3} \code{30} \code{300} \code{3001} \code{301} \code{3010} \code{3011} \code{302} \code{3021} \code{303} \code{3031} \code{304} \code{3041} \code{305} \code{3051} \code{306} \code{3061} \code{307} \code{3071} \code{308} \code{3081} \code{309} \code{3091} \code{31} \code{310} \code{3100} \code{3101} \code{311} \code{3110} \code{3111} \code{312} \code{3121} \code{313} \code{3131} \code{314} \code{3141} \code{315} \code{3151} \code{316} \code{3161} \code{317} \code{3171} \code{318} \code{3181} \code{319} \code{3191} \code{32} \code{320} \code{3201} \code{321} \code{3210} \code{3211} \code{322} \code{3221} \code{323} \code{3231} \code{324} \code{3241} \code{325} \code{3251} \code{326} \code{3261} \code{327} \code{3271} \code{328} \code{3281} \code{329} \code{3291} \code{33} \code{330} \code{3301} \code{331} \code{3310} \code{3311} \code{332} \code{3321} \code{333} \code{3331} \code{334} \code{3341} \code{335} \code{3351} \code{336} \code{3361} \code{337} \code{3371} \code{338} \code{3381} \code{339} \code{3391} \code{34} \code{340} \code{3401} \code{341} \code{3410} \code{3411} \code{342} \code{3421} \code{343} \code{3431} \code{344} \code{3441} \code{345} \code{3451} \code{346} \code{3461} \code{347} \code{3471} \code{348} \code{3481} \code{349} \code{3491} \code{35} \code{350} \code{3501} \code{351} \code{3510} \code{3511} \code{352} \code{3521} \code{353} \code{3531} \code{354} \code{3541} \code{355} \code{3551} \code{356} \code{3561} \code{357} \code{3571} \code{358} \code{3581} \code{359} \code{3591} \code{36} \code{360} \code{3601} \code{361} \code{3610} \code{3611} \code{362} \code{3621} \code{363} \code{3631} \code{364} \code{3641} \code{365} \code{3651} \code{366} \code{3661} \code{367} \code{3671} \code{368} \code{3681} \code{369} \code{3691} \code{37} \code{370} \code{3701} \code{371} \code{3710} \code{3711} \code{372} \code{3721} \code{373} \code{3731} \code{374} \code{3741} \code{375} \code{3751} \code{376} \code{3761} \code{377} \code{3771} \code{378} \code{3781} \code{379} \code{3791} \code{38} \code{380} \code{3801} \code{381} \code{3810} \code{3811} \code{382} \code{3821} \code{383} \code{3831} \code{384} \code{3841} \code{385} \code{3851} \code{386} \code{3861} \code{387} \code{3871} \code{388} \code{3881} \code{389} \code{3891} \code{39} \code{390} \code{3901} \code{391} \code{3910} \code{3911} \code{392} \code{3921} \code{393} \code{3931} \code{394} \code{3941} \code{395} \code{3951} \code{396} \code{3961} \code{397} \code{3971} \code{398} \code{3981} \code{399} \code{3991} \code{4} \code{40} \code{400} \code{4001} \code{401} \code{4010} \code{4011} \code{402} \code{4021} \code{403} \code{4031} \code{404} \code{4041} \code{405} \code{4051} \code{406} \code{4061} \code{407} \code{4071} \code{408} \code{4081} \code{409} \code{4091} \code{41} \code{410} \code{4100} \code{4101} \code{411} \code{4110} \code{4111} \code{412} \code{4121} \code{413} \code{4131} \code{414} \code{4141} \code{415} \code{4151} \code{416} \code{4161} \code{417} \code{4171} \code{418} \code{4181} \code{419} \code{4191} \code{42} \code{420} \code{4201} \code{421} \code{4210} \code{4211} \code{422} \code{4221} \code{423} \code{4231} \code{424} \code{4241} \code{425} \code{4251} \code{426} \code{4261} \code{427} \code{4271} \code{428} \code{4281} \code{429} \code{4291} \code{43} \code{430} \code{4301} \code{431} \code{4310} \code{4311} \code{432} \code{4321} \code{433} \code{4331} \code{434} \code{4341} \code{435} \code{4351} \code{436} \code{4361} \code{437} \code{4371} \code{438} \code{4381} \code{439} \code{4391} \code{44} \code{440} \code{4401} \code{441} \code{4410} \code{4411} \code{442} \code{4421} \code{443} \code{4431} \code{444} \code{4441} \code{445} \code{4451} \code{446} \code{4461} \code{447} \code{4471} \code{448} \code{4481} \code{449} \code{4491} \code{45} \code{450} \code{4501} \code{451} \code{4510} \code{4511} \code{452} \code{4521} \code{453} \code{4531} \code{454} \code{4541} \code{455} \code{4551} \code{456} \code{4561} \code{457} \code{4571} \code{458} \code{4581} \code{459} \code{4591} \code{46} \code{460} \code{4601} \code{461} \code{4610} \code{4611} \code{462} \code{4621} \code{463} \code{4631} \code{464} \code{4641} \code{465} \code{4651} \code{466} \code{4661} \code{467} \code{4671} \code{468} \code{4681} \code{469} \code{4691} \code{47} \code{470} \code{4701} \code{471} \code{4710} \code{4711} \code{472} \code{4721} \code{473} \code{4731} \code{474} \code{4741} \code{475} \code{4751} \code{476} \code{4761} \code{477} \code{4771} \code{478} \code{4781} \code{479} \code{4791} \code{48} \code{480} \code{4801} \code{481} \code{4810} \code{4811} \code{482} \code{4821} \code{483} \code{4831} \code{484} \code{4841} \code{485} \code{4851} \code{486} \code{4861} \code{487} \code{4871} \code{488} \code{4881} \code{489} \code{4891} \code{49} \code{490} \code{4901} \code{491} \code{4910} \code{4911} \code{492} \code{4921} \code{493} \code{4931} \code{494} \code{4941} \code{495} \code{4951} \code{496} \code{4961} \code{497} \code{4971} \code{498} \code{4981} \code{499} \code{4991} \code{5} \code{50} \code{500} \code{5001} \code{501} \code{5010} \code{5011} \code{502} \code{5021} \code{503} \code{5031} \code{504} \code{5041} \code{505} \code{5051} \code{506} \code{5061} \code{507} \code{5071} \code{508} \code{5081} \code{509} \code{5091} \code{51} \code{510} \code{5100} \code{5101} \code{511} \code{5110} \code{5111} \code{512} \code{5121} \code{513} \code{5131} \code{514} \code{5141} \code{515} \code{5151} \code{516} \code{5161} \code{517} \code{5171} \code{518} \code{5181} \code{519} \code{5191} \code{52} \code{520} \code{5201} \code{521} \code{5210} \code{5211} \code{522} \code{5221} \code{523} \code{5231} \code{524} \code{5241} \code{525} \code{5251} \code{526} \code{5261} \code{527} \code{5271} \code{528} \code{5281} \code{529} \code{5291} \code{53} \code{530} \code{5301} \code{531} \code{5310} \code{5311} \code{532} \code{5321} \code{533} \code{5331} \code{534} \code{5341} \code{535} \code{5351} \code{536} \code{5361} \code{537} \code{5371} \code{538} \code{5381} \code{539} \code{5391} \code{54} \code{540} \code{5401} \code{541} \code{5410} \code{5411} \code{542} \code{5421} \code{543} \code{5431} \code{544} \code{5441} \code{545} \code{5451} \code{546} \code{5461} \code{547} \code{5471} \code{548} \code{5481} \code{549} \code{5491} \code{55} \code{550} \code{5501} \code{551} \code{5510} \code{5511} \code{552} \code{5521} \code{553} \code{5531} \code{554} \code{5541} \code{555} \code{5551} \code{556} \code{5561} \code{557} \code{5571} \code{558} \code{5581} \code{559} \code{5591} \code{56} \code{560} \code{5601} \code{561} \code{5610} \code{5611} \code{562} \code{5621} \code{563} \code{5631} \code{564} \code{5641} \code{565} \code{5651} \code{566} \code{5661} \code{567} \code{5671} \code{568} \code{5681} \code{569} \code{5691} \code{57} \code{570} \code{5701} \code{571} \code{5710} \code{5711} \code{572} \code{5721} \code{573} \code{5731} \code{574} \code{5741} \code{575} \code{5751} \code{576} \code{5761} \code{577} \code{5771} \code{578} \code{5781} \code{579} \code{5791} \code{58} \code{580} \code{5801} \code{581} \code{5810} \code{5811} \code{582} \code{5821} \code{583} \code{5831} \code{584} \code{5841} \code{585} \code{5851} \code{586} \code{5861} \code{587} \code{5871} \code{588} \code{5881} \code{589} \code{5891} \code{59} \code{590} \code{5901} \code{591} \code{5910} \code{5911} \code{592} \code{5921} \code{593} \code{5931} \code{594} \code{5941} \code{595} \code{5951} \code{596} \code{5961} \code{597} \code{5971} \code{598} \code{5981} \code{599} \code{5991} \code{6} \code{60} \code{600} \code{6001} \code{601} \code{6010} \code{6011} \code{602} \code{6021} \code{603} \code{6031} \code{604} \code{6041} \code{605} \code{6051} \code{606} \code{6061} \code{607} \code{6071} \code{608} \code{6081} \code{609} \code{6091} \code{61} \code{610} \code{6100} \code{6101} \code{611} \code{6110} \code{6111} \code{612} \code{6121} \code{613} \code{6131} \code{614} \code{6141} \code{615} \code{6151} \code{616} \code{6161} \code{617} \code{6171} \code{618} \code{6181} \code{619} \code{6191} \code{62} \code{620} \code{6201} \code{621} \code{6210} \code{6211} \code{622} \code{6221} \code{623} \code{6231} \code{624} \code{6241} \code{625} \code{6251} \code{626} \code{6261} \code{627} \code{6271} \code{628} \code{6281} \code{629} \code{6291} \code{63} \code{630} \code{6301} \code{631} \code{6310} \code{6311} \code{632} \code{6321} \code{633} \code{6331} \code{634} \code{6341} \code{635} \code{6351} \code{636} \code{6361} \code{637} \code{6371} \code{638} \code{6381} \code{639} \code{6391} \code{64} \code{640} \code{6401} \code{641} \code{6410} \code{6411} \code{642} \code{6421} \code{643} \code{6431} \code{644} \code{6441} \code{645} \code{6451} \code{646} \code{6461} \code{647} \code{6471} \code{648} \code{6481} \code{649} \code{6491} \code{65} \code{650} \code{6501} \code{651} \code{6510} \code{6511} \code{652} \code{6521} \code{653} \code{6531} \code{654} \code{6541} \code{655} \code{6551} \code{656} \code{6561} \code{657} \code{6571} \code{658} \code{6581} \code{659} \code{6591} \code{66} \code{660} \code{6601} \code{661} \code{6610} \code{6611} \code{662} \code{6621} \code{663} \code{6631} \code{664} \code{6641} \code{665} \code{6651} \code{666} \code{6661} \code{667} \code{6671} \code{668} \code{6681} \code{669} \code{6691} \code{67} \code{670} \code{6701} \code{671} \code{6710} \code{6711} \code{672} \code{6721} \code{673} \code{6731} \code{674} \code{6741} \code{675} \code{6751} \code{676} \code{6761} \code{677} \code{6771} \code{678} \code{6781} \code{679} \code{6791} \code{68} \code{680} \code{6801} \code{681} \code{6810} \code{6811} \code{682} \code{6821} \code{683} \code{6831} \code{684} \code{6841} \code{685} \code{6851} \code{686} \code{6861} \code{687} \code{6871} \code{688} \code{6881} \code{689} \code{6891} \code{69} \code{690} \code{6901} \code{691} \code{6910} \code{6911} \code{692} \code{6921} \code{693} \code{6931} \code{694} \code{6941} \code{695} \code{6951} \code{696} \code{6961} \code{697} \code{6971} \code{698} \code{6981} \code{699} \code{6991} \code{7} \code{70} \code{700} \code{7001} \code{701} \code{7010} \code{7011} \code{702} \code{7021} \code{703} \code{7031} \code{704} \code{7041} \code{705} \code{7051} \code{706} \code{7061} \code{707} \code{7071} \code{708} \code{7081} \code{709} \code{7091} \code{71} \code{710} \code{7101} \code{711} \code{7110} \code{7111} \code{712} \code{7121} \code{713} \code{7131} \code{714} \code{7141} \code{715} \code{7151} \code{716} \code{7161} \code{717} \code{7171} \code{718} \code{7181} \code{719} \code{7191} \code{72} \code{720} \code{7201} \code{721} \code{7210} \code{722} \code{723} \code{724} \code{725} \code{726} \code{727} \code{728} \code{729} \code{73} \code{731} \code{74} \code{741} \code{75} \code{751} \code{76} \code{761} \code{77} \code{771} \code{78} \code{781} \code{79} \code{791} \code{8} \code{80} \code{801} \code{81} \code{810} \code{811} \code{82} \code{821} \code{83} \code{831} \code{84} \code{841} \code{85} \code{851} \code{86} \code{861} \code{87} \code{871} \code{88} \code{881} \code{89} \code{891} \code{9} \code{90} \code{901} \code{91} \code{910} \code{911} \code{92} \code{921} \code{93} \code{931} \code{94} \code{941} \code{95} \code{951} \code{96} \code{961} \code{97} \code{971} \code{98} \code{981} \code{99} \code{991}}
    \item{\code{Remun}}{an ordered factor with levels \code{R1} < \code{R2} < \code{R3} < \code{R4} < \code{R5} < \code{R6} < \code{R7} < \code{R8} < \code{R9} < \code{R10}}
    \item{\code{Resrch}}{a factor with levels \code{Res0} \code{Res1} \code{Res2} \code{Res3} \code{Res4} \code{Res5} \code{Res6} \code{Res7}}
    \item{\code{Rank}}{a factor with levels \code{Rnk1} \code{Rnk2} \code{Rnk3} \code{Rnk4} \code{Rnk5}}
    \item{\code{Age}}{a factor with levels \code{A1} \code{A2} \code{A3} \code{A4} \code{A5} \code{A6} \code{A7}}
    \item{\code{Gender}}{a factor with levels \code{Female} \code{Male}}
    \item{\code{PQual}}{a numeric vector}
    \item{\code{AQual}}{a factor with levels \code{Q1} \code{Q2} \code{Q3} \code{Q4} \code{Q5} \code{Q6} \code{Q7} \code{Q8} \code{Q9}}
    \item{\code{Faclty}}{a factor with levels \code{F1} \code{F2} \code{F3} \code{F4} \code{F5} \code{F6} \code{F7} \code{F8} \code{F9}}
    \item{\code{Year}}{a numeric vector}
    \item{\code{Group}}{a factor with levels \code{Female: 2002} \code{Female: 2005} \code{Male: 2002} \code{Male: 2005}}
  }
}

\keyword{datasets}
