\name{Remuneration.data.genbipl.2002}
\alias{Remuneration.data.genbipl.2002}
\docType{data}
\title{
THE 2002 REMUNERATION DATA SET: SOME VARIABLES CONTINUOUS; OTHERS  CATEGORICAL
}
\description{
See Understanding Biplots
}
\usage{data(Remuneration.data.genbipl.2002)}
\format{
  A data frame with 728 observations on the following 7 variables.
  \describe{
    \item{\code{Remun}}{a numeric vector}
    \item{\code{Resrch}}{a numeric vector}
    \item{\code{Age}}{a numeric vector}
    \item{\code{AQual}}{a numeric vector}
    \item{\code{Rank}}{a factor with levels \code{R1} \code{R2} \code{R3} \code{R4} \code{R5}}
    \item{\code{Gender}}{a factor with levels \code{Female} \code{Male}}
    \item{\code{Faclty}}{a factor with levels \code{F1} \code{F2} \code{F3} \code{F4} \code{F5} \code{F6} \code{F7} \code{F8} \code{F9}}
  }
}

\keyword{datasets}
