\name{Remuneration.data}
\alias{Remuneration.data}
\docType{data}
\title{
THE REMUNERATION DATA SET
}
\description{
See Understanding Biplots
}
\usage{data(Remuneration.data)}
\format{
  A data frame with 1448 observations on the following 11 variables.
  \describe{
    \item{\code{ID}}{a numeric vector}
    \item{\code{Remun}}{a numeric vector}
    \item{\code{Resrch}}{a numeric vector}
    \item{\code{Rank}}{a numeric vector}
    \item{\code{Age}}{a numeric vector}
    \item{\code{Gender}}{a character vector}
    \item{\code{PQual}}{a numeric vector}
    \item{\code{AQual}}{a numeric vector}
    \item{\code{Faclty}}{a numeric vector}
    \item{\code{Year}}{a numeric vector}
    \item{\code{Group}}{a factor with levels \code{Female: 2002} \code{Female: 2005} \code{Male: 2002} \code{Male: 2005}}
  }
}

\keyword{datasets}
