\name{Rika.data}
\alias{Rika.data}
\docType{data}
\title{
THE 2006 EDUCATION DATA SET
}
\description{
See Understanding Biplots.
}
\usage{data(Rika.data)}
\format{
  A data frame with 24470 observations on the following 20 variables.
  \describe{
    \item{\code{UqNr}}{a numeric vector}
    \item{\code{PersonNr}}{a numeric vector}
    \item{\code{Prov}}{a numeric vector}
    \item{\code{dc}}{a character vector}
    \item{\code{Gender}}{a numeric vector}
    \item{\code{Age}}{a numeric vector}
    \item{\code{Race}}{a numeric vector}
    \item{\code{RelatToHead}}{a numeric vector}
    \item{\code{MaritalS}}{a numeric vector}
    \item{\code{ExpTransp}}{a numeric vector}
    \item{\code{ExpHouse}}{a numeric vector}
    \item{\code{ExpCloth}}{a numeric vector}
    \item{\code{ExpFood}}{a numeric vector}
    \item{\code{ExpPersApp}}{a numeric vector}
    \item{\code{ExpOthr}}{a numeric vector}
    \item{\code{EducY}}{a numeric vector}
    \item{\code{TotExpR}}{a numeric vector}
    \item{\code{percExpG}}{a numeric vector}
    \item{\code{ParentEd}}{a numeric vector}
    \item{\code{ParentG}}{a numeric vector}
  }
}

\keyword{datasets}
