\name{Rothkopf.vowels}
\alias{Rothkopf.vowels}
\docType{data}
\title{
THE ROTHKOPF VOWELS DATA SET
}
\description{
See Understanding Biplots
}
\usage{data(Rothkopf.vowels)}
\format{
  The format is:
 int [1:5, 1:5] 92 6 64 7 14 3 97 10 5 4 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:5] "a" "e" "i" "o" ...
  ..$ : chr [1:5] "a" "e" "i" "o" ...
}

\keyword{datasets}
