\name{RSACrime.data}
\alias{RSACrime.data}
\docType{data}
\title{
THE SOUTH AFRICAN CRIME DATA SET
}
\description{
See Understanding Biplots.
}
\usage{data(RSACrime.data)}
\format{
  The format is:
List of 14
 $ Arson  : int [1:10, 1:7] 1438 377 1815 1866 909 612 536 265 921 8739 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:10] "EasternCape" "FreeState" "Gauteng" "KwaZulu-Natal" ...
  .. ..$ : chr [1:7] "X2001.2002" "X2002.2003" "X2003.2004" "X2004.2005" ...
 $ AssGBH : int [1:10, 1:7] 41119 17485 57672 34307 23131 20320 20420 13076 36482 264012 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:10] "EasternCape" "FreeState" "Gauteng" "KwaZulu-Natal" ...
  .. ..$ : chr [1:7] "X2001.2002" "X2002.2003" "X2003.2004" "X2004.2005" ...
 $ AtMur  : int [1:10, 1:7] 3777 1482 8189 7258 1580 1698 1653 1677 3979 31293 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:10] "EasternCape" "FreeState" "Gauteng" "KwaZulu-Natal" ...
  .. ..$ : chr [1:7] "X2001.2002" "X2002.2003" "X2003.2004" "X2004.2005" ...
 $ BurNRes: int [1:10, 1:7] 9513 6299 19445 14054 6719 5724 6356 3118 15886 87114 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:10] "EasternCape" "FreeState" "Gauteng" "KwaZulu-Natal" ...
  .. ..$ : chr [1:7] "X2001.2002" "X2002.2003" "X2003.2004" "X2004.2005" ...
 $ BurRes : int [1:10, 1:7] 35603 19072 84610 49051 15362 21307 16853 6778 54021 302657 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:10] "EasternCape" "FreeState" "Gauteng" "KwaZulu-Natal" ...
  .. ..$ : chr [1:7] "X2001.2002" "X2002.2003" "X2003.2004" "X2004.2005" ...
 $ CarJack: int [1:10, 1:7] 818 145 9872 2845 347 673 344 9 793 15846 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:10] "EasternCape" "FreeState" "Gauteng" "KwaZulu-Natal" ...
  .. ..$ : chr [1:7] "X2001.2002" "X2002.2003" "X2003.2004" "X2004.2005" ...
 $ ComAss : int [1:10, 1:7] 29607 25307 62844 36097 21306 13012 16661 9300 47752 261886 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:10] "EasternCape" "FreeState" "Gauteng" "KwaZulu-Natal" ...
  .. ..$ : chr [1:7] "X2001.2002" "X2002.2003" "X2003.2004" "X2004.2005" ...
 $ ComRob : int [1:10, 1:7] 10313 4453 29867 14708 4458 3978 5702 2326 14400 90205 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:10] "EasternCape" "FreeState" "Gauteng" "KwaZulu-Natal" ...
  .. ..$ : chr [1:7] "X2001.2002" "X2002.2003" "X2003.2004" "X2004.2005" ...
 $ DrugRel: int [1:10, 1:7] 7253 4572 9095 10374 1785 1683 2742 1967 13429 52900 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:10] "EasternCape" "FreeState" "Gauteng" "KwaZulu-Natal" ...
  .. ..$ : chr [1:7] "X2001.2002" "X2002.2003" "X2003.2004" "X2004.2005" ...
 $ IndAss : int [1:10, 1:7] 532 327 1244 864 177 190 314 209 1782 5639 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:10] "EasternCape" "FreeState" "Gauteng" "KwaZulu-Natal" ...
  .. ..$ : chr [1:7] "X2001" "X2002" "X2003" "X2004" ...
 $ Murd   : int [1:10, 1:7] 3483 926 4937 5441 803 958 935 475 3447 21405 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:10] "EasternCape" "FreeState" "Gauteng" "KwaZulu-Natal" ...
  .. ..$ : chr [1:7] "X2001.2002" "X2002.2003" "X2003.2004" "X2004.2005" ...
 $ PubViol: int [1:10, 1:7] 205 49 118 84 30 76 23 74 248 907 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:10] "EasternCape" "FreeState" "Gauteng" "KwaZulu-Natal" ...
  .. ..$ : chr [1:7] "X2001.2002" "X2002.2003" "X2003.2004" "X2004.2005" ...
 $ Rape   : int [1:10, 1:7] 5133 2886 10201 7071 3550 2828 3065 1233 5115 41082 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:10] "EasternCape" "FreeState" "Gauteng" "KwaZulu-Natal" ...
  .. ..$ : chr [1:7] "X2001" "X2002" "X2003" "X2004" ...
 $ RobAC  : int [1:10, 1:7] 8514 2534 56167 24184 3254 5518 3743 522 12300 116736 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:10] "EasternCape" "FreeState" "Gauteng" "KwaZulu-Natal" ...
  .. ..$ : chr [1:7] "X2001.2002" "X2002.2003" "X2003.2004" "X2004.2005" ...
}

\keyword{datasets}
