\name{SACrime02.data}
\alias{SACrime02.data}
\docType{data}
\title{
THE 2002SOUTH AFRICAN CRIME DATA SET
}
\description{
See Understanding Biplots.
}
\usage{data(SACrime02.data)}
\format{
  The format is:
 int [1:9, 1:14] 1438 377 1815 1866 909 612 536 265 921 41119 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:9] "ECpe" "FrSt" "Gaut" "KZN" ...
  ..$ : chr [1:14] "Arsn" "AGBH" "AtMr" "BNRs" ...
}

\keyword{datasets}
