\name{SACrime08.data}
\alias{SACrime08.data}
\docType{data}
\title{
THE 2008 SOUTH AFRICAN CRIME DATA SET
}
\description{
See Understanding Biplots.
}
\usage{data(SACrime08.data)}
\format{
  The format is:
 int [1:9, 1:14] 1235 432 1824 1322 573 588 624 169 629 34479 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:9] "ECpe" "FrSt" "Gaut" "KZN" ...
  ..$ : chr [1:14] "Arsn" "AGBH" "AtMr" "BNRs" ...
}

\keyword{datasets}
