\name{SACrimeGaut.data}
\alias{SACrimeGaut.data}
\docType{data}
\title{
THE GAUTENG CRIME DATA SET
}
\description{
See Understanding Biplots.
}
\usage{data(SACrimeGaut.data)}
\format{
  The format is:
 int [1:7, 1:14] 1815 1977 1983 1949 1916 2039 1824 57672 57838 55867 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:7] "02" "03" "04" "05" ...
  ..$ : chr [1:14] "Arsn" "AGBH" "AtMr" "BNRs" ...
}

\keyword{datasets}
