\name{SACrimeWCpe.data}
\alias{SACrimeWCpe.data}
\docType{data}
\title{
THE WESTERN CAPE CRIME DATA SET
}
\description{
See Understanding Biplots.
}
\usage{data(SACrimeWCpe.data)}
\format{
  The format is:
 int [1:7, 1:14] 921 960 967 720 595 625 629 36482 36122 36912 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:7] "02" "03" "04" "05" ...
  ..$ : chr [1:14] "Arsn" "AGBH" "AtMr" "BNRs" ...
}

\keyword{datasets}
