\name{SAPPI.data}
\alias{SAPPI.data}
\docType{data}
\title{
THE SOUTH AFRICAN PINUS DATA SET
}
\description{
See Understanding Biplots.
}
\usage{data(SAPPI.data)}
\format{
  A data frame with 36 observations on the following 9 variables.
  \describe{
    \item{\code{species}}{a factor with levels \code{P.ell} \code{P.kes} \code{P.max} \code{P.pat} \code{P.tae}}
    \item{\code{tot.yield}}{a numeric vector}
    \item{\code{alkali}}{a numeric vector}
    \item{\code{density}}{a numeric vector}
    \item{\code{TEA}}{a numeric vector}
    \item{\code{tensile}}{a numeric vector}
    \item{\code{tear}}{a numeric vector}
    \item{\code{burst}}{a numeric vector}
    \item{\code{growth}}{a numeric vector}
  }
}

\keyword{datasets}
