\name{soils.data}
\alias{soils.data}
\docType{data}
\title{
THE SOILS DATA SET
}
\description{
See Understanding Biplots.
}
\usage{data(soils.data)}
\format{
  The format is:
 num [1:15, 1:12] 0.1174 0.0845 0.1734 0.1631 -0.112 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:15] "a" "b" "c" "d" ...
  ..$ : chr [1:12] "V1" "V2" "V3" "V4" ...
}

\keyword{datasets}
