\name{Table.1.2.data}
\alias{Table.1.2.data}
\docType{data}
\title{
THE TABLE 1.2 DATA SET
}
\description{
See Understanding Biplots.
}
\usage{data(Table.1.2.data)}
\format{
  The format is:
 num [1:6, 1:4] 47.1 48.6 59.5 63.6 71.2 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:6] "90" "130" "140" "150" ...
  ..$ : chr [1:4] "Amea" "Edun" "Egran" "Emac"
}

\keyword{datasets}
