\name{Table.3.3.data}
\alias{Table.3.3.data}
\docType{data}
\title{
THE TABLE 3.3 DATA SET
}
\description{
See Understanding Biplots.
}
\usage{data(Table.3.3.data)}
\format{
  A data frame with 25 observations on the following 4 variables.
  \describe{
    \item{\code{SampleNo}}{a numeric vector}
    \item{\code{X}}{a numeric vector}
    \item{\code{Y}}{a numeric vector}
    \item{\code{Z}}{a numeric vector}
  }
}

\keyword{datasets}
