\name{Table.3.4.data}
\alias{Table.3.4.data}
\docType{data}
\title{
THE TABLE 3.4 DATA SET
}
\description{
See Understanding Biplots.
}
\usage{data(Table.3.4.data)}
\format{
  The format is:
 num [1:21, 1:6] 1.255 1.085 0.445 0.539 0.614 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:21] "a" "b" "c" "d" ...
  ..$ : chr [1:6] "V1" "V2" "V3" "V4" ...
}

\keyword{datasets}
