\name{Table.7.2.data}
\alias{Table.7.2.data}
\docType{data}
\title{
THE TABLE 7.2 DATA SET
}
\description{
See Understanding Biplots.
}
\usage{data(Table.7.2.data)}
\format{
  The format is:
 int [1:4, 1:6] 12 43 83 73 10 37 79 67 8 71 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:4] "Prod.A" "Prod.B" "Prod.C" "Prod.D"
  ..$ : chr [1:6] "Attr.1" "Attr.2" "Attr.3" "Attr.4" ...
}

\keyword{datasets}
