\name{UBbipl-package}
\alias{UBbipl-package}
\alias{UBbipl}
\docType{package}
\title{
DATA SETS AND FUNCTIONS IN UNDERSTANDING BIPLOTS INCLUDING
SUBSEQUENTLY WRITTEN R FUNCTIONS
}
\description{
All the data sets and functions described in
 Understanding Biplots including some functions written subsequently

}
\details{
\tabular{ll}{
Package: \tab UBbipl\cr
Type: \tab Package\cr
Version: \tab 3.0.4\cr
Date: \tab 2013-10-13\cr
License: \tab AGPL-3\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
The usage of the package is described in detail in Understanding Biplots.
}
\references{
Gower, JC, Lubbe, S and Le Roux, NJ. 2011. Understanding Biplots. Wiley: Chichester
}
\seealso{
UBFigs
}
