\name{University.carillon.data}
\alias{University.carillon.data}
\docType{data}
\title{
THE UNIVERSITY CARILLON DATA SET
}
\description{
See Understanding Biplots.
}
\usage{data(University.carillon.data)}
\format{
  A data frame with 63 observations on the following 20 variables.
  \describe{
    \item{\code{No}}{a numeric vector}
    \item{\code{Founder}}{a factor with levels \code{Eisjb} \code{GilJoh}}
    \item{\code{Year}}{a numeric vector}
    \item{\code{Diameter}}{a numeric vector}
    \item{\code{Weight}}{a numeric vector}
    \item{\code{PITCH}}{a numeric vector}
    \item{\code{KEYNOTE}}{a numeric vector}
    \item{\code{PRIME}}{a numeric vector}
    \item{\code{THIRD}}{a numeric vector}
    \item{\code{QUINT}}{a numeric vector}
    \item{\code{OCTAVE}}{a numeric vector}
    \item{\code{LogDia}}{a numeric vector}
    \item{\code{LogW}}{a numeric vector}
    \item{\code{LogRatio}}{a numeric vector}
    \item{\code{DEVPITCH}}{a numeric vector}
    \item{\code{DEVKEYNOTE}}{a numeric vector}
    \item{\code{DEVPRIME}}{a numeric vector}
    \item{\code{DEVTHIRD}}{a numeric vector}
    \item{\code{DEVQUINT}}{a numeric vector}
    \item{\code{DEVOCTAVE}}{a numeric vector}
  }
}

\keyword{datasets}
