\name{VAR95.data}
\alias{VAR95.data}
\docType{data}
\title{
THE VALUE-AT-RISK DATA SET
}
\description{
See Understanding Biplots.
}
\usage{data(VAR95.data)}
\format{
  A data frame with 20 observations on the following 8 variables.
  \describe{
    \item{\code{Day}}{a numeric vector}
    \item{\code{CM}}{a numeric vector}
    \item{\code{IRD}}{a numeric vector}
    \item{\code{MM}}{a numeric vector}
    \item{\code{ALCO}}{a numeric vector}
    \item{\code{SE}}{a numeric vector}
    \item{\code{EDSA}}{a numeric vector}
    \item{\code{EDM}}{a numeric vector}
  }
}

\keyword{datasets}
