\name{wheat.data}
\alias{wheat.data}
\docType{data}
\title{
THE WHEAT DATA SET
}
\description{
See Understanding Biplots.
}
\usage{data(wheat.data)}
\format{
  The format is:
 num [1:14, 1:12] 321 411 317 429 364 464 408 434 419 492 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:14] "Cra.L" "Cra.H" "Beg.L" "Beg.H" ...
  ..$ : chr [1:12] "Cap" "Ran" "Hun" "Tem" ...
}

\keyword{datasets}
